+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Unreleased
 
+### Added
+* Added additional session statuses for audio device I/O timeouts.
+
 ### Fixed
 * [Demo] Fixed the remote video render delay
+
 
 ## [0.17.9] - 2022-12-02